Add type to ApplicationCommandInteractionDataOption via new ApplicationCommandOptionTypeTrait
