+++ Tests/Objects/Slash/ApplicationCommandInteractionDataOptionTest.php
@@ -3,6 +3,7 @@
 namespace Bytes\DiscordResponseBundle\Tests\Objects\Slash;
 
 use Bytes\Common\Faker\TestFakerTrait;
+use Bytes\DiscordResponseBundle\Enums\ApplicationCommandOptionType;
 use Bytes\DiscordResponseBundle\Objects\Slash\ApplicationCommandInteractionDataOption;
 use Bytes\Tests\Common\DataProvider\BooleanProviderTrait;
 use Bytes\Tests\Common\DataProvider\NullProviderTrait;
@@ -94,12 +95,18 @@ class ApplicationCommandInteractionDataOptionTest extends TestCase
      * @param $value
      * @param $options
      */
-    public function testCreate($name, $value, $options)
+    public function testCreate($name, $type, $value, $options)
     {
-        $option = ApplicationCommandInteractionDataOption::create($name, $value, $options);
+        $option = ApplicationCommandInteractionDataOption::create($name, $type, $value, $options);
         $this->assertEquals($name, $option->getName());
         $this->assertEquals($value, $option->getValue());
         $this->assertEquals($options, $option->getOptions());
+
+        if($type instanceof ApplicationCommandOptionType)
+        {
+            $type = $type->value;
+        }
+        $this->assertEquals($type, $option->getType());
     }
 
     /**
@@ -110,11 +117,13 @@ class ApplicationCommandInteractionDataOptionTest extends TestCase
     {
         $this->setupFaker();
         foreach ([$this->faker->word(), $this->faker->userName(), $this->faker->numberBetween(), $this->faker->boolean(), null] as $value) {
-            yield ['name' => $this->faker->word(), 'value' => $value, 'options' => null];
-            yield ['name' => $this->faker->word(), 'value' => $value, 'options' => []];
-            foreach (range(1, 5) as $optionsRange) {
-                $options[] = new ApplicationCommandInteractionDataOption();
-                yield ['name' => $this->faker->word(), 'value' => $value, 'options' => $options];
+            foreach ([ApplicationCommandOptionType::string(), ApplicationCommandOptionType::string()->value, null] as $type) {
+                yield ['name' => $this->faker->word(), 'type' => $type, 'value' => $value, 'options' => null];
+                yield ['name' => $this->faker->word(), 'type' => $type, 'value' => $value, 'options' => []];
+                foreach (range(1, 5) as $optionsRange) {
+                    $options[] = new ApplicationCommandInteractionDataOption();
+                    yield ['name' => $this->faker->word(), 'type' => $type, 'value' => $value, 'options' => $options];
+                }
             }
         }
     }
